+++ extensions/word-search/src/word-rhyme-search.tsx
@@ -1,10 +1,7 @@
 import { SearchType } from "@/types";
 
-import { useClipboardSetting } from "@/hooks/use-settings";
-
 import SearchResults from "@/components/SearchResults";
 
 export default function SearchRhyme() {
-  const useClipboard = useClipboardSetting<Preferences.WordRhymeSearch>();
-  return SearchResults(SearchType.RHYME, "Search for rhymes", useClipboard);
+  return SearchResults(SearchType.RHYME, "Search for rhymes");
 }