+++ extensions/word-search/src/word-synonym-search.tsx
@@ -1,10 +1,7 @@
 import { SearchType } from "@/types";
 
-import { useClipboardSetting } from "@/hooks/use-settings";
-
 import SearchResults from "@/components/SearchResults";
 
 export default function SearchSynonym() {
-  const useClipboard = useClipboardSetting<Preferences.WordSynonymSearch>();
-  return SearchResults(SearchType.SYNONYM, "Search for synonyms", useClipboard);
+  return SearchResults(SearchType.SYNONYM, "Search for synonyms");
 }